+++ src/errors.ts
@@ -1,0 +1,29 @@
+export class SDJWTError extends Error {
+  constructor(message: any) {
+    super(message);
+  }
+}
+
+export class DecodeJWTError extends SDJWTError {
+  constructor(message: any) {
+    super(message);
+  }
+}
+
+export class IssueSDJWTError extends SDJWTError {
+  constructor(message: any) {
+    super(message);
+  }
+}
+
+export class VerifySDJWTError extends SDJWTError {
+  constructor(message: any) {
+    super(message);
+  }
+}
+
+export class PackSDJWTError extends SDJWTError {
+  constructor(message: any) {
+    super(message);
+  }
+}