+++ src/types.ts
@@ -69,7 +69,6 @@ export interface IssueSDJWTOptions {
 export interface VerifySdJwtOptions {
   kb?: {
     verifier?: KeyBindingVerifier;
-    skipCheck?: boolean;
   };
 }
 /**