+++ CHANGELOG.md
@@ -10,5 +10,7 @@ and this project (loosely) adheres to [Semantic Versioning](https://semver.org/s
 Initial version
 
 ### Changed
+- removed kb jwt payload checks
+- added error types
 - added simple demo scripts
 - add `.js` file extensions to all imports for ESM compatibility